Refactor fingerprint code into its own directory
