+++ packages/cli/src/cli.js
@@ -7,7 +7,6 @@
 /* eslint-disable max-classes-per-file */
 
 const yargs = require('yargs');
-const chokidar = require('chokidar');
 const { diffLines } = require('diff');
 const yaml = require('js-yaml');
 const { promises: fsp, readFileSync } = require('fs');
@@ -16,99 +15,15 @@ const process = require('process');
 const readline = require('readline');
 const { join } = require('path');
 const { algorithms, canonicalize } = require('./fingerprint');
-const { verbose, listAppMapFiles, loadAppMap } = require('./utils');
+const { verbose, loadAppMap } = require('./utils');
 const appMapCatalog = require('./appMapCatalog');
-const FingerprintQueue = require('./fingerprintQueue');
+const FingerprintDirectoryCommand = require('./fingerprint/fingerprintDirectoryCommand');
+const FingerprintWatchCommand = require('./fingerprint/fingerprintWatchCommand');
 const Depends = require('./depends');
 const FindCodeObjects = require('./search/findCodeObjects');
 const FindEvents = require('./search/findEvents');
 const FunctionStats = require('./functionStats');
 const Inspect = require('./inspect');
-
-class FingerprintDirectoryCommand {
-  constructor(directory) {
-    this.directory = directory;
-    this.print = false;
-  }
-
-  setPrint(print) {
-    this.print = print;
-    return this;
-  }
-
-  async execute() {
-    if (verbose()) {
-      console.warn(`Fingerprinting appmaps in ${this.directory}`);
-    }
-
-    const fpQueue = new FingerprintQueue();
-    await this.files(fpQueue.push.bind(fpQueue));
-    await fpQueue.process();
-  }
-
-  async files(fn) {
-    return listAppMapFiles(this.directory, fn);
-  }
-}
-
-class FingerprintWatchCommand {
-  constructor(directory) {
-    this.directory = directory;
-    this.print = false;
-    this.numProcessed = 0;
-  }
-
-  setPrint(print) {
-    this.print = print;
-    return this;
-  }
-
-  execute() {
-    if (verbose()) {
-      console.warn(`Watching appmaps in ${this.directory}`);
-    }
-
-    this.fpQueue = new FingerprintQueue();
-    this.fpQueue.setCounterFn(() => {
-      this.numProcessed += 1;
-    });
-    this.fpQueue.process();
-    const watcher = chokidar.watch(`${this.directory}/**/*.appmap.json`, {
-      ignoreInitial: true,
-    });
-    watcher
-      .on('add', this.added.bind(this))
-      .on('change', this.changed.bind(this))
-      .on('unlink', this.removed.bind(this));
-  }
-
-  added(file) {
-    if (verbose()) {
-      console.warn(`AppMap added: ${file}`);
-    }
-    this.enqueue(file);
-  }
-
-  changed(file) {
-    if (verbose()) {
-      console.warn(`AppMap changed: ${file}`);
-    }
-    this.enqueue(file);
-  }
-
-  // eslint-disable-next-line class-methods-use-this
-  removed(file) {
-    console.warn(`TODO: AppMap removed: ${file}`);
-  }
-
-  enqueue(file) {
-    // This shouldn't be necessary, but it's passing through the wrong file names.
-    if (!file.includes('.appmap.json')) {
-      return;
-    }
-    this.fpQueue.push(file);
-  }
-}
 
 class DiffCommand {
   baseDir(dir) {
